+++ workout_diary/src/entities/workout.py
@@ -1,5 +1,5 @@
 from entities.user import User
-
+import datetime
 
 class Workout:
     """Class for a one workout object. 
@@ -27,20 +27,32 @@ class Workout:
                 - VO2 Maximum    
     """
 
-    def __init__(self, name: str, repetition: bool, workout_type: str, workout_id: None):
+    def __init__(self, name: str, user: User, date_and_time: datetime.date, repetition: bool, workout_type: bool, sets: str, details: str):
 
-        self.workout_name = name
-        self.workout_repetition = repetition  # Weekly: True, Monthly: False
-        self.workout_type = workout_type
-        self.workout_details_weight_lifting = {}
-        self.workout_sets_weight_lifting = {}
-        self.workout_details_cardio = {}
-        self.workout_id = workout_id
+        self.workout_name = name #String: workout's name
+        self.user = user #User(username, password)
+        self.date_and_time = date_and_time #datetime.date(YY, MM, DD) 
+        self.repetition = repetition #Boolean: True means weekly, False monthly
+        self.workout_type = workout_type #Boolean: True means weight lift, False cardio
+        self.sets = sets #String
+        self.details = details #String
 
-    def add_weight_lifting_set(self, name, repetitions, weight):
-        "Method adds new set for a wight lift workout"
-        self.workout_sets_weight_lifting[name] = (repetitions, weight)
+    def get_workout_name(self):
+        return self.workout_name
+
+    def get_workout_user(self):
+        return self.user
+    
+    def get_workout_date_and_time(self):
+        return self.date_and_time
+
+    def get_workout_sets(self):
+        return self.sets
+    
+    def get_workout_repetition(self):
+        return self.repetition
+
+    def get_workout_details(self):
+        return self.details
 
 
-if __name__ == "__main__":
-    treeni = Workout("pena", True, "weight_lift")