+++ workout_diary/src/ui/login_window.py
@@ -1,12 +1,16 @@
-from tkinter import ttk, constants
+from tkinter import ttk, constants, StringVar
+from services.user_service import user_service
+from entities.user import User
 
 class LoginWindow:
-    def __init__(self, root, handle_exit, handle_create_user, handle_login_user):
+    def __init__(self, root, handle_exit, handle_create_user, show_main_view):
         self._root = root
         self._handle_exit = handle_exit
         self._handle_create_user = handle_create_user
-        self._handle_login_user = handle_login_user
+        self._show_main_view = show_main_view
         self._frame = None
+        self._error_message = None
+        self._error_label = None
 
         self._initialize()
 
@@ -16,17 +20,30 @@ class LoginWindow:
     def destroy(self):
         self._frame.destroy()
 
+    def _hide_error(self):
+        self._error_label.grid_remove()
+
     def _initialize(self):
         self._frame = ttk.Frame(master=self._root)
         heading_label = ttk.Label(master=self._frame, text="Welcome to Workout Diary!")
         
+        self._error_message = StringVar(self._frame)
+
+        self._error_label = ttk.Label(
+            master=self._frame,
+            textvariable=self._error_message,
+            foreground="red"
+        )
+
+        self._error_label.grid(padx=5, pady=5)
+
         login_label = ttk.Label(master=self._frame, text="Login:")
 
-        username_label = ttk.Label(master=self._frame, text="Username:")
+        username_label = ttk.Label(master=self._frame, text="Enter username:")
         self._username_entry = ttk.Entry(master=self._frame)
 
-        password_label = ttk.Label(master=self._frame, text="Password:")
-        self._login_password = ttk.Entry(master=self._frame)
+        password_label = ttk.Label(master=self._frame, text="Enter password:")
+        self._password_entry = ttk.Entry(master=self._frame)
 
         button_stop_application = ttk.Button(master=self._frame, 
                             text='Stop application',
@@ -51,7 +68,7 @@ class LoginWindow:
         self._username_entry.grid(row=2, column=1, sticky=(constants.E, constants.W), padx=5, pady=5)
 
         password_label.grid(row=3, column=0, padx=5, pady=5)
-        self._login_password.grid(row=3, column=1, sticky=(constants.E, constants.W), padx=5, pady=5)
+        self._password_entry.grid(row=3, column=1, sticky=(constants.E, constants.W), padx=5, pady=5)
 
         button_login_user.grid(row=4, column=0, columnspan=2, padx=5, pady=5)
 
@@ -60,3 +77,24 @@ class LoginWindow:
         button_stop_application.grid(row=6, column=0, columnspan=2, padx=5, pady=5)
 
         self._frame.grid_columnconfigure(1, weight = 1, minsize=400)
+
+        self._hide_error()
+
+    def _handle_login_user(self):
+        
+        self._login_username = self._username_entry.get()
+        self._login_password = self._password_entry.get()
+
+        try:
+            user_service.login_user(self._login_username, self._login_password)
+            self._show_main_view
+        except:
+            self._show_error("Invalid username or password")
+
+    def _show_error(self, message):
+        self._error_message.set(message)
+        self._error_label.grid()
+
+        #user_logged_in = user_service.login_user(self._login_username, self._login_password_hashed)
+        #user_logged_in = user_service.login_user(self._login_username, self._login_password)
+